#157030

Related work items: #157030

+++ Logitude/Logitude.BL/CommonDataModel/EntityLists/HTSCodeList.cs
@@ -15,8 +15,8 @@ namespace Logitude.BL.CommonDataModel.EntityLists
         public bool ApprovedByCustomer { get; set; }
         public bool InActive { get; set; }
         public int? LineNumber { get; set; }
-        public decimal? VATPercentage { get; set; }
-        public decimal? DutiesPercentage { get; set; }
+        public double? VATPercentage { get; set; }
+        public double? DutiesPercentage { get; set; }
         public string OtherDuties { get; set; }
         public string Remarks { get; set; }
     }
